made test discovery more flexible
Tests can now be in folders instead of only in root of test project.
Test project namespace does not have to match test project folder name anymore.

This was achieved by getting a hold on the type of the testclass, finding the root project folder of the testclass en then find a file that has the same name as the testclass somewhere in the directory structure of the test project